+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - A `Pagerfanta\Exception\InvalidArgumentException` is now raised when the `$routeGenerator` is not a callable, as of 3.0 all methods will typehint the requirement
 - Deprecated `Pagerfanta\Adapter\MandangoAdapter`, the dependent package is abandoned
 - Deprecated `Pagerfanta\Adapter\PropelAdapter` and `Pagerfanta\Adapter\Propel2Adapter`
+- Deprecated `Pagerfanta\Adapter\MongoAdapter` as it relies on the older `ext/mongo`